refactore code #62

+++ snugh/lecture/views.py
@@ -1,10 +1,10 @@
 from django.shortcuts import get_object_or_404
-from django.db.models import Q
 from django.db.models import Case, When
 from django.db import transaction
-from rest_framework import status, viewsets, filters
+from rest_framework import status, viewsets
 from rest_framework.response import Response
 from rest_framework.decorators import action
+from rest_framework.permissions import IsAuthenticated
 from lecture.models import * 
 from lecture.serializers import *
 from user.models import *
@@ -13,8 +13,9 @@ from django.core.paginator import Paginator
 from django.db.models.functions import Length
 from django.db.models import F
 from snugh.permissions import IsOwnerOrCreateReadOnly
-from snugh.exceptions import FieldError, NotFound, NotOwner
+from snugh.exceptions import NotOwner, DuplicationError
 from lecture.utils import add_credits, subtract_credits, add_semester_credits, sub_semester_credits
+from .const import *
 
 
 class PlanViewSet(viewsets.GenericViewSet):
@@ -24,7 +25,7 @@ class PlanViewSet(viewsets.GenericViewSet):
 
     queryset = Plan.objects.all()
     serializer_class = PlanSerializer 
-    #permission_classes = [IsOwnerOrCreateReadOnly]
+    permission_classes = [IsOwnerOrCreateReadOnly]
 
     # POST /plan
     @transaction.atomic
@@ -80,7 +81,7 @@ class PlanViewSet(viewsets.GenericViewSet):
 
     # 강의구분 자동계산
     # PUT /plan/:planId/calculate
-    @action(detail=True, methods=['GET'])
+    @action(detail=True, methods=['PUT'])
     @transaction.atomic
     def calculate(self, request, pk=None):
         """Calculate credits"""
@@ -96,6 +97,8 @@ class PlanViewSet(viewsets.GenericViewSet):
             ).get(id=pk)
             
         user = plan.user
+        if request.user != user:
+            raise NotOwner()
 
         semesters = plan.semester.all()
         planmajors = plan.planmajor.all()
@@ -127,18 +130,19 @@ class PlanViewSet(viewsets.GenericViewSet):
                     semester = sub_semester_credits(semesterlecture, semester)
                     lecture = semesterlecture.lecture
 
-                    if semesterlecture.lecture_type != SemesterLecture.GENERAL:
+                    if semesterlecture.lecture_type != GENERAL:
                         major_count = 0
                         std = user.userprofile.entrance_year
                         majorlectures = lecture.majorlecture.all()
                         for major in tmp_majors:
                             if major_count > 1:
                                 break
-                            candidate_majorlectures = majorlectures.filter(major=major,
-                                                                                start_year__lte=std,
-                                                                                end_year__gte=std)\
-                                .exclude(lecture_type__in=[MajorLecture.GENERAL, MajorLecture.GENERAL_ELECTIVE])\
-                                .order_by('-lecture_type')
+                            candidate_majorlectures = majorlectures.filter(
+                                major=major,
+                                start_year__lte=std,
+                                end_year__gte=std)\
+                            .exclude(lecture_type__in=[GENERAL, GENERAL_ELECTIVE])\
+                            .order_by('-lecture_type')
 
                             if candidate_majorlectures.exists():
                                 candidate_majorlecture = candidate_majorlectures.first()
@@ -160,11 +164,12 @@ class PlanViewSet(viewsets.GenericViewSet):
                                 if major_count > 1:
                                     break
 
-                                candidate_majorlectures = lecture.majorlecture.filter(major=major,
-                                                                                    start_year__lte=std,
-                                                                                    end_year__gte=std)\
-                                    .exclude(lecture_type__in=[MajorLecture.GENERAL, MajorLecture.GENERAL_ELECTIVE])\
-                                    .order_by('-lecture_type')
+                                candidate_majorlectures = lecture.majorlecture.filter(
+                                    major=major,
+                                    start_year__lte=std,
+                                    end_year__gte=std)\
+                                .exclude(lecture_type__in=[GENERAL, GENERAL_ELECTIVE])\
+                                .order_by('-lecture_type')
 
                                 if candidate_majorlectures.exists() != 0:
                                     candidate_majorlecture = candidate_majorlectures.first()
@@ -178,13 +183,13 @@ class PlanViewSet(viewsets.GenericViewSet):
                                     major_count += 1
 
                         if major_count == 1:
-                            semesterlecture.lecture_type2 = SemesterLecture.NONE
+                            semesterlecture.lecture_type2 = NONE
                             semesterlecture.recognized_major2 = none_major
                         elif major_count == 0:
-                            semesterlecture.lecture_type = SemesterLecture.GENERAL_ELECTIVE
-                            semesterlecture.lecture_type1 = SemesterLecture.GENERAL_ELECTIVE
+                            semesterlecture.lecture_type = GENERAL_ELECTIVE
+                            semesterlecture.lecture_type1 = GENERAL_ELECTIVE
                             semesterlecture.recognized_major1 = none_major
-                            semesterlecture.lecture_type2 = SemesterLecture.NONE
+                            semesterlecture.lecture_type2 = NONE
                             semesterlecture.recognized_major2 = none_major
 
                     lecturecredits = lecture.lecturecredit.filter(start_year__lte=semester.year,
@@ -274,11 +279,12 @@ class SemesterViewSet(viewsets.GenericViewSet):
     """
     queryset = Semester.objects.all()
     serializer_class = SemesterSerializer
-    permission_classes = [IsOwnerOrCreateReadOnly]
+    permission_classes = [IsAuthenticated]
 
     # POST /semester
     @transaction.atomic
-    def create(self, request): 
+    def create(self, request):
+        """Create new semester"""
         data = request.data
         serializer = self.get_serializer(data=data)
         serializer.is_valid(raise_exception=True)
@@ -288,24 +294,26 @@ class SemesterViewSet(viewsets.GenericViewSet):
     # PUT /semester/:semesterId
     @transaction.atomic
     def update(self, request, pk=None):
-        user = request.user
+        """Update semester"""
         data = request.data
 
         semester = self.get_object()
         plan = semester.plan
+        if plan.user != request.user:
+            raise NotOwner()
         year = data.get('year')
         semester_type = data.get('semester_type')
         is_complete = data.get('is_complete')
 
         if (not is_complete) and (not year) and (not semester_type):
-            return Response({"error": "body is empty"}, status=status.HTTP_403_FORBIDDEN)
+            raise FieldError("body is empty")
 
         if (not is_complete) == ((not year) and (not semester_type)):
-            return Response({"error": "is_complete should be none"}, status=status.HTTP_403_FORBIDDEN)
+            raise FieldError("is_complete should be none")
 
         if not is_complete:
-            if Semester.objects.filter(plan=plan, year=year, semester_type=semester_type).exists():
-                return Response({"error": "semester already_exist"}, status=status.HTTP_403_FORBIDDEN)
+            if plan.semester.filter(plan=plan, year=year, semester_type=semester_type).exists():
+                raise DuplicationError("semester already_exist")
 
         serializer = self.get_serializer(semester, data=data, partial=True)
         serializer.is_valid(raise_exception=True)
@@ -314,9 +322,11 @@ class SemesterViewSet(viewsets.GenericViewSet):
     
     # DEL /semester/:semesterId
     def destroy(self, request, pk=None):
-        user = request.user
+        """Destroy semester"""
         semester = self.get_object()
         plan = semester.plan
+        if plan.user != request.user:
+            raise NotOwner()
         semester.delete()
         # TODO: update plan info when delete semester
         # update_plan_info(plan=plan)
@@ -324,11 +334,10 @@ class SemesterViewSet(viewsets.GenericViewSet):
     
     # GET /semester/:semesterId
     def retrieve(self, request, pk=None):
-        user = request.user
+        """Retrieve semester"""
         semester = self.get_object() 
         serializer = SemesterSerializer(semester)
         return Response(serializer.data, status=status.HTTP_200_OK) 
-
 
 class LectureViewSet(viewsets.GenericViewSet):
     queryset = SemesterLecture.objects.all()
